Map render bug fixes
