+++ backend/https/routes/index.js
@@ -1,16 +1,21 @@
 const { Router } = require("express");
-const { User, Avatar, Map } = require("../db/db.js");
-const { SignupSchema, SigninSchema, AvatarSchema, MapSchema, DeleteMapSchema, DeleteAvatarSchema } = require("../types/index.js");
+const { User } = require("../db/db.js");
+const { SignupSchema, SigninSchema, MapSchema } = require("../types/index.js");
 const { z } = require("zod");
 const jwt = require("jsonwebtoken");
-const authMiddleware = require("../middleware/auth.js");
 const { JWT_PASSWORD } = require("../../config/config.js");
-const { upload } = require("../middleware/upload.js");
 const cors = require("cors");
 const express = require("express");
+const path = require("path");
 
 const app = express();
 app.use(cors());
+
+// const absolutePath = "D:\source\RandD\2d\Office-Metaverse\backend\uploads";
+const uploadsPath = path.resolve(__dirname, '../uploads');
+app.use('/uploads', express.static(uploadsPath));
+// app.use("/uploads", express.static(path.join(__dirname, "https", "uploads")));
+console.log("Image path: " + uploadsPath);
 
 const router = Router();
 
@@ -77,155 +82,10 @@ router.post("/signin", async (req, res) => {
     }
 });
 
-router.post("/createAvatar", authMiddleware, upload.single('avatarImage'), async (req, res) => {
-    try {
-        req.body.avatarId = parseInt(req.body.avatarId, 10);
+// Avatar endpoints
 
-        const validatedBody = AvatarSchema.parse(req.body);
-        const { avatarId, avatarName } = validatedBody;
 
-        if (!req.file) {
-            return res.status(400).send({ msg: "No avatar uploaded." });
-        }
+// Map endpoints
 
-        const avatarImagePath = req.file.path;
-
-        const isAvatar = await Avatar.findOne({ avatarId });
-        if (!isAvatar) {
-            await Avatar.create({ avatarId, avatarName, imagePath: avatarImagePath });
-            return res.status(200).send({
-                msg: `Avatar ${avatarName} created successfully.`
-            });
-        }
-
-        res.status(400).send({
-            msg: `Avatar already exists.`
-        });
-
-    } catch (error) {
-        if (error instanceof z.ZodError) {
-            return res.status(400).send({
-                msg: "Invalid input data.",
-                errors: error.errors
-            });
-        }
-
-        res.status(500).json({
-            msg: "Something went wrong. Avatar not created."
-        });
-    }
-});
-
-router.get("/fetchAvatar", async (req, res) => {
-    const avatars = await Avatar.find();
-
-    if (!avatars) {
-        res.status(404).send({
-            msg: "No Avatars found."
-        })
-    }
-
-    res.status(200).send({
-        msg: "Avatars fetched successfully.",
-        avatars
-    })
-});
-
-router.delete("/avatars/:avatarId", authMiddleware, async (req, res) => {
-    try {
-        const { avatarId } = req.params;
-        const deleteAvatar = await Avatar.findOneAndDelete({ avatarId });
-
-        if (deleteAvatar) {
-            res.status(200).send({
-                msg: `Avatar ${avatarId} deleted successfully.`
-            })
-        } else {
-            res.status(404).send({
-                msg: `Avatar ${avatarId} not found.`
-            });
-        }
-    } catch (error) {
-        res.status(500).json({
-            msg: "Something went wrong. Couldn't delete the avatar."
-        });
-    }
-});
-
-router.post("/createMap", authMiddleware, upload.single('mapImage'), async (req, res) => {
-    try {
-        req.body.mapId = parseInt(req.body.mapId, 10);
-
-        const validatedBody = MapSchema.parse(req.body);
-        const { mapId, mapName } = validatedBody;
-
-        if (!req.file) {
-            return res.status(400).send({ msg: "No map uploaded." });
-        }
-
-        const mapImagePath = req.file.path;
-
-        const isMap = await Map.findOne({ mapId });
-        if (!isMap) {
-            await Map.create({ mapId, mapName, imagePath: mapImagePath });
-            return res.status(200).send({
-                msg: `Map ${mapName} created successfully.`,
-                mapImagePath
-            });
-        }
-
-        res.status(400).send({
-            msg: `Map already exists.`
-        });
-
-    } catch (error) {
-        if (error instanceof z.ZodError) {
-            return res.status(400).send({
-                msg: "Invalid input data.",
-                errors: error.errors
-            });
-        }
-
-        res.status(500).json({
-            msg: "Something went wrong. Map not created."
-        });
-    }
-});
-
-router.get("/fetchMap", async (req, res) => {
-    const maps = await Map.find();
-
-    if (!maps) {
-        res.status(404).send({
-            msg: "No Maps found."
-        })
-    }
-
-    res.status(200).send({
-        msg: "Maps fetched successfully.",
-        maps
-    })
-});
-
-router.delete("/maps/:mapId", authMiddleware, async (req, res) => {
-    try {
-        const { mapId } = req.params;
-        const deleteMap = await Map.findOneAndDelete({ mapId });
-
-        if (deleteMap) {
-            res.status(200).send({
-                msg: `Map ${mapId} deleted successfully.`
-            })
-        } else {
-            res.status(404).send({
-                msg: `Map ${mapId} not found.`
-            });
-        }
-    } catch (error) {
-        res.status(500).json({
-            msg: "Something went wrong. Couldn't delete the map."
-        });
-    }
-});
 
 module.exports = { router };
