+++ package.json
@@ -6,6 +6,8 @@
     "start": "concurrently \"npm run start:server-https\" \"npm run start:server-ws\" \"npm run start:frontend\""
   },
   "dependencies": {
+    "cors": "^2.8.5",
+    "path": "^0.12.7",
     "zod": "^3.23.8"
   },
   "devDependencies": {
